doc, refactor
