+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/callable/function/Function.java
@@ -25,8 +25,7 @@ import org.enso.interpreter.runtime.Context;
 import org.enso.interpreter.runtime.callable.CallerInfo;
 import org.enso.interpreter.runtime.callable.argument.ArgumentDefinition;
 import org.enso.interpreter.runtime.callable.argument.Thunk;
-import org.enso.interpreter.runtime.data.EmptyState;
-import org.enso.interpreter.runtime.data.SmallMap;
+import org.enso.interpreter.runtime.state.data.EmptyMap;
 import org.enso.interpreter.runtime.data.Vector;
 import org.enso.interpreter.runtime.type.Types;
 import org.enso.polyglot.MethodNames;
@@ -200,7 +199,7 @@ public final class Function implements TruffleObject {
         Object[] arguments,
         @Cached InteropApplicationNode interopApplicationNode,
         @CachedContext(Language.class) Context context) {
-      return interopApplicationNode.execute(function, EmptyState.create(), arguments);
+      return interopApplicationNode.execute(function, EmptyMap.create(), arguments);
     }
   }
